Fix currency
