+++ example/lib/navigation/tab_navigator_home.dart
@@ -10,7 +10,7 @@ import '../screens/themes/theme_options_screen.dart';
 import '../screens/themes/theme_colors_screen.dart';
 import '../screens/themes/theme_fonts_screen.dart';
 import '../screens/textfields/number_screen.dart';
-import '../screens/currency_screen.dart';
+import '../screens/currency/currency_screen.dart';
 import '../screens/radio_button_screen.dart';
 import '../screens/buttons/buttons_ghost_screen.dart';
 import '../screens/buttons/buttons_primary_screen.dart';
@@ -44,10 +44,10 @@ import '../screens/tabs/informer_tabs_screen.dart';
 import '../screens/tabs/tabs_standard_screen.dart';
 import '../screens/informers/toast_screen.dart';
 import '../screens/checkbox_screen.dart';
-import '../screens/currency_icon_screen.dart';
-import '../screens/currency_default_screen.dart';
-import '../screens/currency_flags_screen.dart';
-import '../screens/currency_icon_flags_screen.dart';
+import '../screens/currency/currency_icon_screen.dart';
+import '../screens/currency/currency_default_screen.dart';
+import '../screens/currency/currency_flags_screen.dart';
+import '../screens/currency/currency_icon_flags_screen.dart';
 import '../screens/textfields/pin_code_screen.dart';
 import '../screens/textfields/double_textfield_screen.dart';
 import '../screens/cells/base_cell_screen.dart';
