+++ example/lib/screens/currency/currency_flags_screen.dart
@@ -1,7 +1,7 @@
 import 'package:admiralui_flutter/admiralui_flutter.dart';
 import 'package:admiralui_flutter/layout/layout_grid.dart';
 import 'package:flutter/material.dart';
-import '../navigation/tab_navigator_home.dart';
+import '../../navigation/tab_navigator_home.dart';
 
 class CurrencyFlagsScreen extends StatelessWidget {
   const CurrencyFlagsScreen({
@@ -66,14 +66,15 @@ class CurrencyFlagsScreen extends StatelessWidget {
               'Валюта',
               'Купить',
               'Продать',
+              isTextSpacingEnabled: true,
             ),
             SizedBox(
               height: LayoutGrid.doubleModule,
             ),
             Currency(
               'USD',
-              '89',
-              '90',
+              '68,65',
+              '67,88',
               Image(
                 image: AssetImage('assets/currency/usa.png'),
                 fit: BoxFit.cover,
@@ -87,8 +88,8 @@ class CurrencyFlagsScreen extends StatelessWidget {
             ),
             Currency(
               'EUR',
-              '89',
-              '90',
+              '68,65',
+              '67,88',
               Image(
                 image: AssetImage('assets/currency/european.png'),
                 fit: BoxFit.cover,
@@ -102,8 +103,8 @@ class CurrencyFlagsScreen extends StatelessWidget {
             ),
             Currency(
               'GBP',
-              '89',
-              '90',
+              '68,65',
+              '67,88',
               Image(
                 image: AssetImage('assets/currency/britain.png'),
                 fit: BoxFit.cover,
@@ -117,8 +118,8 @@ class CurrencyFlagsScreen extends StatelessWidget {
             ),
             Currency(
               'CNY',
-              '89',
-              '90',
+              '68,65',
+              '67,88',
               Image(
                 image: AssetImage('assets/currency/china.png'),
                 fit: BoxFit.cover,
