+++ example/lib/screens/currency/currency_screen.dart
@@ -1,7 +1,7 @@
 import 'package:admiralui_flutter/admiralui_flutter.dart';
 import 'package:admiralui_flutter/layout/layout_grid.dart';
 import 'package:flutter/material.dart';
-import '../navigation/tab_navigator_home.dart';
+import '../../navigation/tab_navigator_home.dart';
 
 class CurrencyScreen extends StatelessWidget {
   const CurrencyScreen({
